1. Two Sum

+++ Algorithms/two-sum.cpp
@@ -1,16 +1,17 @@
 //1. Two Sum
 //https://leetcode.com/problems/two-sum/
-// https://discuss.leetcode.com/topic/3294/accepted-c-o-n-solution/1
+// https://discuss.leetcode.com/topic/3294/accepted-c-o-n-solution/
 #include <iostream>
 #include <vector>
 #include <unordered_map>
 using namespace std;
+// BEGIN: https://discuss.leetcode.com/topic/3294/accepted-c-o-n-solution/
 class Solution {
 public:
 	vector<int> twoSum(const vector<int>& nums, const int& target) {
 		unordered_map<int, int> hashmap;
 		vector<int> result(2, 0);
-		for (int i = 0; i < nums.size(); ++i) {
+		for (size_t i = 0; i < nums.size(); ++i) {
 			int rest = target - nums[i];
 			if (hashmap.count(rest)) {
 				result[0] = hashmap[rest];
@@ -22,6 +23,7 @@ public:
 		return result;
 	}
 };
+// END: https://discuss.leetcode.com/topic/3294/accepted-c-o-n-solution/
 int main(int argc, char** argv) {
 	Solution solution;
 	vector<int> nums = {2, 7, 11, 15};
